refactor(media): movies & shows into own folders
